refactored code for modular testing
