+++ DeepDream3D/ModelDefinition/utils.py
