+++ DeepDream3D/preprocessing/3_gather_img.py
