+++ DeepDream3D/preprocessing/4_gather_all_vox_img_train.py
