+++ DeepDream3D/preprocessing/4_gather_vox_img_train.py
