+++ DeepDream3D/scratchpad.py
