+++ bash_run_scripts/test_ae.sh
@@ -1,5 +1,5 @@
 python ../Insight_Project_Framework/main.py \
-                --ae \
+                --ae True\
                 --data_dir /data/IM-NET-pytorch/data/all_vox256_img \
                 --checkpoint_dir /data/IM-NET-pytorch/checkpoint \
                 --sample_dir /data/IM-NET-pytorch/data/all_vox256_img \
