+++ tests/ModelDefinition/modelAE_DD_test.py
@@ -1,0 +1,4 @@
+import hdf5
+
+
+from DeepDream3D.ModelDefinition import modelAE_DD
